ci: allow GORM implementation build triggers to continue on error
The `hibernate6` repository currently does not have GitHub workflows enabled, causing the job to fail and preventing subsequent steps from running.

This commit configures `continue-on-error: true` at the job level to ensure the workflow does not fail if a trigger fails, and at the step level to ensure subsequent steps are still triggered.