feat(grails-geb): Read GebConfig.groovy to allow overriding RemoteWebDriver (#14987)
* refactor(grails-geb): Read GebConfig.groovy to allow overriding RemoteWebDriver, Close #14801

I hereby lived under the assumption that this does not need to be as dynamic as the other grails-geb/connection settings (which can be defined per Class using Annotation/Interface), and just needs to be project-configureable.
Though that may raise the question of some users why FileDetector was made configurable in this way too.

This now should also open up all other gebconfig functionality.

Holding out with writing Documentation until feedback on idea's was provided. Maybe I also did something in an geb-unwanted way.
_Also possibly currently trying to make a diagram to better wrap my head around grails-geb's moving parts and necessities for future encounters._

<details>
<summary>Unrelevant Thoughts</summary>
I had originally tried to allow the use of normal Drivers and then extracting their capabilities,
but as far as I was able to jungle through this is not possible.
One option there would've been to intercept e.g. ChromeOptions construction when calling closure,
or another hacking interception way, but i chose not to.

I also had the idea to introduce a new GebConfig key, like other *-geb extensions do (e.g. grailsGeb, ).
</details>

* test(geb): extract `GebConfig` testing to separate project

Do a separate project for testing using default `GebConfig` file
so we don't pollute the other tests.

* refactor: prefer `.with` over `.tap`

`.tap` returns the receiver, which suggests fluent chaining.
Since the return value is unused, use `.with`.
Behavior is unchanged.

* chore: cleanup

* fix: add back options to default chrome driver

---------

Co-authored-by: JonasPammer <jonas.pammer@nesevo.com>
Co-authored-by: Mattias Reichel <mattias.reichel@gmail.com>
Co-authored-by: Mattias Reichel <matrei@apache.org>